Bug 1859391 - Update cookie banner settings page to only support private mode.
