+++ fenix/app/metrics.yaml
@@ -8389,27 +8389,10 @@ autoplay:
         - SitePermissions
 
 cookie_banners:
-  visited_setting:
-    type: event
-    description: A user visited the cookie banner handling screen
-    bugs:
-      - https://bugzilla.mozilla.org/show_bug.cgi?id=1796146
-    data_reviews:
-      - https://github.com/mozilla-mobile/fenix/pull/27561
-      - https://github.com/mozilla-mobile/firefox-android/pull/2597
-      - https://github.com/mozilla-mobile/firefox-android/pull/4039
-    data_sensitivity:
-      - interaction
-    notification_emails:
-      - android-probes@mozilla.com
-    expires: never
-    metadata:
-      tags:
-        - Privacy&Security
-  setting_changed:
+  setting_changed_pmb:
     type: event
     description: |
-      A user changed their setting.
+      A user changed their setting in private mode.
     extra_keys:
       cookie_banner_setting:
         description: |
@@ -8475,80 +8458,6 @@ cookie_banners:
       - https://bugzilla.mozilla.org/show_bug.cgi?id=1797577
     data_reviews:
       - https://github.com/mozilla-mobile/fenix/pull/28044#issuecomment-1334548056
-      - https://github.com/mozilla-mobile/firefox-android/pull/2597
-      - https://github.com/mozilla-mobile/firefox-android/pull/4039
-    data_sensitivity:
-      - interaction
-    notification_emails:
-      - android-probes@mozilla.com
-    expires: never
-    metadata:
-      tags:
-        - Privacy&Security
-  visited_re_engagement_dialog:
-    type: event
-    description: An user visited the cookie banner re-engagement dialog
-    bugs:
-      - https://bugzilla.mozilla.org/show_bug.cgi?id=1797593
-    data_reviews:
-      - https://github.com/mozilla-mobile/fenix/pull/28405#issuecomment-1372489596
-      - https://github.com/mozilla-mobile/firefox-android/pull/2597
-      - https://github.com/mozilla-mobile/firefox-android/pull/4039
-    data_sensitivity:
-      - interaction
-    notification_emails:
-      - android-probes@mozilla.com
-    expires: never
-    metadata:
-      tags:
-        - Privacy&Security
-  opt_out_re_engagement_dialog:
-    type: event
-    description: |
-      An user opt out the cookie banner re-engagement
-      dialog by clicking the X button
-    bugs:
-      - https://bugzilla.mozilla.org/show_bug.cgi?id=1797593
-    data_reviews:
-      - https://github.com/mozilla-mobile/fenix/pull/28405#issuecomment-1372489596
-      - https://github.com/mozilla-mobile/firefox-android/pull/2597
-      - https://github.com/mozilla-mobile/firefox-android/pull/4039
-    data_sensitivity:
-      - interaction
-    notification_emails:
-      - android-probes@mozilla.com
-    expires: never
-    metadata:
-      tags:
-        - Privacy&Security
-  not_now_re_engagement_dialog:
-    type: event
-    description: |
-      An user clicked the not now button on
-      the cookie banner re-engagement dialog
-    bugs:
-      - https://bugzilla.mozilla.org/show_bug.cgi?id=1797593
-    data_reviews:
-      - https://github.com/mozilla-mobile/fenix/pull/28405#issuecomment-1372489596
-      - https://github.com/mozilla-mobile/firefox-android/pull/2597
-      - https://github.com/mozilla-mobile/firefox-android/pull/4039
-    data_sensitivity:
-      - interaction
-    notification_emails:
-      - android-probes@mozilla.com
-    expires: never
-    metadata:
-      tags:
-        - Privacy&Security
-  allow_re_engagement_dialog:
-    type: event
-    description: |
-      An user clicked the dismiss banner button
-      on the cookie banner re-engagement dialog
-    bugs:
-      - https://bugzilla.mozilla.org/show_bug.cgi?id=1797593
-    data_reviews:
-      - https://github.com/mozilla-mobile/fenix/pull/28405#issuecomment-1372489596
       - https://github.com/mozilla-mobile/firefox-android/pull/2597
       - https://github.com/mozilla-mobile/firefox-android/pull/4039
     data_sensitivity:
