+++ fenix/app/src/androidTest/java/org/mozilla/fenix/ui/SettingsPrivacyTest.kt
@@ -55,8 +55,7 @@ class SettingsPrivacyTest {
             verifyPrivateBrowsingButton()
             verifyHTTPSOnlyModeButton()
             verifySettingsOptionSummary("HTTPS-Only Mode", "Off")
-            verifyCookieBannerReductionButton()
-            verifySettingsOptionSummary("Cookie banner reduction", "Off")
+            verifySettingsOptionSummary("Cookie Banner Blocker in private browsing", "")
             verifyEnhancedTrackingProtectionButton()
             verifySettingsOptionSummary("Enhanced Tracking Protection", "Standard")
             verifySitePermissionsButton()
