+++ fenix/app/src/main/java/org/mozilla/fenix/BrowserDirection.kt
@@ -26,7 +26,6 @@ enum class BrowserDirection(@IdRes val fragmentId: Int) {
     FromAbout(R.id.aboutFragment),
     FromTrackingProtection(R.id.trackingProtectionFragment),
     FromHttpsOnlyMode(R.id.httpsOnlyFragment),
-    FromCookieBanner(R.id.cookieBannerFragment),
     FromTrackingProtectionDialog(R.id.trackingProtectionPanelDialogFragment),
     FromSavedLoginsFragment(R.id.savedLoginsFragment),
     FromAddNewDeviceFragment(R.id.addNewDeviceFragment),
