+++ fenix/app/src/main/java/org/mozilla/fenix/settings/SupportUtils.kt
@@ -55,7 +55,6 @@ object SupportUtils {
         SMARTBLOCK("smartblock-enhanced-tracking-protection"),
         SPONSOR_PRIVACY("sponsor-privacy"),
         HTTPS_ONLY_MODE("https-only-mode-firefox-android"),
-        COOKIE_BANNER("cookie-banner-reduction-firefox-android"),
         UNSIGNED_ADDONS("unsigned-addons"),
         REVIEW_QUALITY_CHECK("review_checker_mobile"),
         FX_SUGGEST("search-suggestions-firefox"),
