--- fenix/app/src/main/res/xml/cookie_banner_preferences.xml
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?><!-- This Source Code Form is subject to the terms of the Mozilla Public
-   - License, v. 2.0. If a copy of the MPL was not distributed with this
-   - file, You can obtain one at http://mozilla.org/MPL/2.0/. -->
-<androidx.preference.PreferenceScreen xmlns:android="http://schemas.android.com/apk/res/android">
-    <org.mozilla.fenix.settings.CookieBannersSwitchPreference
-        android:defaultValue="false"
-        android:key="@string/pref_key_cookie_banner_v1"
-        android:title="@string/reduce_cookie_banner_option" />
-</androidx.preference.PreferenceScreen>
